+++ irs-api/src/main/java/org/eclipse/tractusx/irs/aaswrapper/submodel/domain/SubmodelFacade.java
@@ -27,6 +27,7 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.irs.component.Relationship;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StopWatch;
 
 /**
  * Public API Facade for submodel domain
@@ -44,8 +45,11 @@ public class SubmodelFacade {
      * @return The Aspect Model for the given submodel
      */
     public List<Relationship> getRelationships(final String submodelEndpointAddress, final RelationshipAspect traversalAspectType) {
+        final StopWatch stopWatch = new StopWatch();
+        stopWatch.start("Get Submodel task for relationships");
         final RelationshipSubmodel relationshipSubmodel = this.submodelClient.getSubmodel(submodelEndpointAddress, traversalAspectType.getSubmodelClazz());
-
+        stopWatch.stop();
+        log.info("Task {} took {} ms for endpoint address: {}", stopWatch.getLastTaskName(), stopWatch.getLastTaskTimeMillis(), submodelEndpointAddress);
         return relationshipSubmodel.asRelationships();
     }
 
@@ -54,7 +58,11 @@ public class SubmodelFacade {
      * @return The Aspect Model as JSON-String for the given submodel
      */
     public String getSubmodelRawPayload(final String submodelEndpointAddress) {
+        final StopWatch stopWatch = new StopWatch();
+        stopWatch.start("Get Submodel raw payload task");
         final String submodel = this.submodelClient.getSubmodel(submodelEndpointAddress);
+        stopWatch.stop();
+        log.info("Task {} took {} ms for endpoint address: {}", stopWatch.getLastTaskName(), stopWatch.getLastTaskTimeMillis(), submodelEndpointAddress);
         log.info("Returning Submodel as String: '{}'", submodel);
         return submodel;
     }