+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added new parameters 'startedOn' and 'jobCompleted' to Job status response 
+
 ### Changed
 - Updated Spring Boot to 2.7.5 and Spring Security (Web and OAuth2 Client) dependencies to 5.7.5 due to CVEs
+- Renamed parameter from 'status' to 'jobState' in Job status response
+- Time to live for finished jobs is now configurable
 
 ## [1.4.0] - 2022-10-28
 ### Added